fix: gate active inventory on provider-current work
